+++ controller/controlplane/controller.go
@@ -423,6 +423,12 @@ func (r *Reconciler) constructControlPlaneManagerConfigOptions(
 		)
 	}
 
+	if cp.Spec.DataPlaneSync != nil {
+		cfgOpts = append(cfgOpts,
+			WithDataPlaneSyncOptions(*cp.Spec.DataPlaneSync),
+		)
+	}
+
 	// If the ControlPlane is owned by a Gateway, we set the Gateway to be the only one to reconcile.
 	if owner, ok := lo.Find(cp.GetOwnerReferences(), func(owner metav1.OwnerReference) bool {
 		return strings.HasPrefix(owner.APIVersion, gatewayv1.GroupName) &&